changed the icons bar to a component
