+++ frontend/HarmonyMobileApp/Components/FilterModal.js
@@ -5,6 +5,7 @@ import {
   StatusBar,
   TextInput,
   TouchableOpacity,
+  ScrollView
 } from "react-native";
 import styles from "../styles";
 import {
@@ -59,10 +60,11 @@ export default function FilterModal({ color, title, ...otherProps }) {
       onBackdropPress={toggleModal}
       animationIn={"slideInRight"}
       animationOut={"slideOutRight"}
-      swipeDirection={["up", "left", "right", "down"]}
+      // swipeDirection={["up", "left", "right", "down"]}
+      // swipeDirection={["left", "right"]}
       onSwipeComplete={toggleModal}
     >
-      <StatusBar hidden={false} />
+      <StatusBar hidden={true} />
 
       <View style={styles.centeredView}>
         <View style={styles.modalView}>
@@ -97,199 +99,200 @@ export default function FilterModal({ color, title, ...otherProps }) {
               />
             </TouchableOpacity>
           </View>
-
-          <View style={styles.filterView}>
-            <View style={[styles.filterLabelRow]}>
-              <Text
-                style={[styles.spaceLeft, styles.TextSmall, { paddingTop: 5 }]}
-              >
-                Sort Pairings
-              </Text>
-              <View style={[styles.spaceRight, { flex: 0 }]}>
-                <AntDesign
-                  name="caretdown"
-                  size={12}
-                  style={{ paddingVertical: "10%", marginLeft: 10 }}
-                  color="#7C7C7C"
-                />
+          <ScrollView style={{ width: "100%", height: "85%" }}>
+            <View style={styles.filterView}>
+              <View style={[styles.filterLabelRow]}>
                 <Text
+                  style={[styles.spaceLeft, styles.TextSmall, { paddingTop: 5 }]}
+                >
+                  Sort Pairings
+                </Text>
+                <View style={[styles.spaceRight, { flex: 0 }]}>
+                  <AntDesign
+                    name="caretdown"
+                    size={12}
+                    style={{ paddingVertical: "10%", marginLeft: 10 }}
+                    color="#7C7C7C"
+                  />
+                  <Text
+                    style={[
+                      styles.TextSmaller,
+                      {
+                        flex: 1,
+                        textAlign: "center",
+                        justifyContent: "center",
+                      },
+                    ]}
+                  >
+                    {sortPairings}
+                  </Text>
+                </View>
+              </View>
+              <View style={styles.pickerView}>
+                <Picker
+                  selectedValue={sortPairings}
+                  style={[styles.TextSmall, { height: 40, width: 300 }]}
+                  onValueChange={(itemValue, itemIndex) => {
+                    setSortPairings(itemValue);
+                    console.log(itemValue);
+                  }}
+                >
+                  <Picker.Item label="Trending" value="Trending" />
+                  <Picker.Item label="Most Liked" value="Most Liked" />
+                  <Picker.Item label="Newest" value="Newest" />
+                  <Picker.Item label="Controversial" value="Controversial" />
+                </Picker>
+              </View>
+            </View>
+            <Text style={{ height: 7 }}></Text>
+
+            <View style={styles.filterView}>
+              <View style={[styles.filterLabelRow]}>
+                <Text style={[styles.spaceLeft, styles.TextSmall]}>Distance</Text>
+              </View>
+              <View
+                style={{
+                  flexDirection: "row",
+                  justifyContent: "center",
+                  marginTop: -10,
+                }}
+              >
+                <Text style={[styles.TextSmall, { marginRight: 4 }]}>0</Text>
+                <Slider
+                  value={locationValue}
+                  step={20}
+                  maximumValue={100}
+                  onValueChange={(value) => (
+                    console.log(value), setLocationValue(value)
+                  )}
+                  style={{ width: "70%" }}
+                  thumbStyle={{
+                    width: 20,
+                    height: 20,
+                    backgroundColor: "grey",
+                  }}
+                />
+                <Text style={[styles.TextSmall, { marginLeft: 4 }]}>100</Text>
+              </View>
+              <View style={{ flexDirection: "row", justifyContent: "center" }}>
+                <TextInput
+                  style={[styles.TextSmall, styles.TextInputStyling]}
+                  value={locationValue}
+                  // when you use to string it causes errors with the slider
+                  onChangeText={(value) => setLocationValue(parseInt(value))}
+                  keyboardType="numeric"
+                  placeholder={locationValue.toString()}
+                  multiline={false}
+                />
+                <Text style={[styles.TextSmaller]}>KM</Text>
+              </View>
+            </View>
+            <Text style={{ height: 7 }}></Text>
+
+            <View style={styles.filterView}>
+              <View style={[styles.filterLabel]}>
+                <View style={{ flexDirection: "row" }}>
+                  <MaterialIcons
+                    name="fastfood"
+                    size={18}
+                    style={{ paddingTop: 3, marginRight: 5 }}
+                    color="black"
+                  />
+                  <Text style={[styles.spaceLeft, styles.TextSmall]}>
+                    Meal Types
+                  </Text>
+                </View>
+                <View style={[styles.filterTagsContainer]}>
+                  <View
+                    style={{
+                      flexDirection: "row",
+                      flexWrap: "wrap",
+                      justifyContent: "space-evenly",
+                    }}
+                  >
+                    {/* Mapping of the tags from the JSON to the components*/}
+                    {filters.mealTypes.map((tag, index) => (
+                      <View key={index}>
+                        <FilterTag color="#FF6347" title={tag} />
+                      </View>
+                    ))}
+                  </View>
+                </View>
+              </View>
+            </View>
+            <Text style={{ height: 7 }}></Text>
+
+            <View style={styles.filterView}>
+              <View style={[styles.filterLabel, { marginBottom: 35 }]}>
+                <View style={{ flexDirection: "row" }}>
+                  <MaterialCommunityIcons
+                    name="cup"
+                    size={18}
+                    style={{ paddingTop: 9, marginRight: 5 }}
+                    color="black"
+                  />
+                  <Text style={[styles.spaceLeft, styles.TextSmall]}>Drinks</Text>
+                </View>
+                <View
                   style={[
-                    styles.TextSmaller,
-                    {
-                      flex: 1,
-                      textAlign: "center",
-                      justifyContent: "center",
-                    },
+                    styles.filterTagsContainer,
+                    { flexDirection: "column" },
                   ]}
                 >
-                  {sortPairings}
-                </Text>
-              </View>
-            </View>
-            <View style={styles.pickerView}>
-              <Picker
-                selectedValue={sortPairings}
-                style={[styles.TextSmall, { height: 40, width: 300 }]}
-                onValueChange={(itemValue, itemIndex) => {
-                  setSortPairings(itemValue);
-                  console.log(itemValue);
-                }}
-              >
-                <Picker.Item label="Trending" value="Trending" />
-                <Picker.Item label="Most Liked" value="Most Liked" />
-                <Picker.Item label="Newest" value="Newest" />
-                <Picker.Item label="Controversial" value="Controversial" />
-              </Picker>
-            </View>
-          </View>
-          <Text style={{ height: 7 }}></Text>
-
-          <View style={styles.filterView}>
-            <View style={[styles.filterLabelRow]}>
-              <Text style={[styles.spaceLeft, styles.TextSmall]}>Distance</Text>
-            </View>
-            <View
-              style={{
-                flexDirection: "row",
-                justifyContent: "center",
-                marginTop: -10,
-              }}
-            >
-              <Text style={[styles.TextSmall, { marginRight: 4 }]}>0</Text>
-              <Slider
-                value={locationValue}
-                step={20}
-                maximumValue={100}
-                onValueChange={(value) => (
-                  console.log(value), setLocationValue(value)
-                )}
-                style={{ width: "70%" }}
-                thumbStyle={{
-                  width: 20,
-                  height: 20,
-                  backgroundColor: "grey",
-                }}
-              />
-              <Text style={[styles.TextSmall, { marginLeft: 4 }]}>100</Text>
-            </View>
-            <View style={{ flexDirection: "row", justifyContent: "center" }}>
-              <TextInput
-                style={[styles.TextSmall, styles.TextInputStyling]}
-                value={locationValue.toString()}
-                onChangeText={(value) => setLocationValue(parseInt(value))}
-                keyboardType="numeric"
-                placeholder={locationValue.toString()}
-                multiline={false}
-              />
-              <Text style={[styles.TextSmaller]}>KM</Text>
-            </View>
-          </View>
-          <Text style={{ height: 7 }}></Text>
-
-          <View style={styles.filterView}>
-            <View style={[styles.filterLabel]}>
-              <View style={{ flexDirection: "row" }}>
-                <MaterialIcons
-                  name="fastfood"
-                  size={18}
-                  style={{ paddingTop: 3, marginRight: 5 }}
-                  color="black"
-                />
-                <Text style={[styles.spaceLeft, styles.TextSmall]}>
-                  Meal Types
-                </Text>
-              </View>
-              <View style={[styles.filterTagsContainer]}>
-                <View
-                  style={{
-                    flexDirection: "row",
-                    flexWrap: "wrap",
-                    justifyContent: "space-evenly",
-                  }}
-                >
-                  {/* Mapping of the tags from the JSON to the components*/}
-                  {filters.mealTypes.map((tag, index) => (
-                    <View key={index}>
-                      <FilterTag color="#FF6347" title={tag} />
-                    </View>
-                  ))}
+                  <View
+                    style={{
+                      flexDirection: "row",
+                      flexWrap: "wrap",
+                      justifyContent: "space-evenly",
+                    }}
+                  >
+                    {/* Mapping of the tags from the JSON to the components*/}
+                    {filters.drinks.map((tag, index) => (
+                      <View key={index}>
+                        <FilterTag color="#1FBFBA" title={tag} />
+                      </View>
+                    ))}
+                  </View>
                 </View>
               </View>
             </View>
-          </View>
-          <Text style={{ height: 7 }}></Text>
+            <Text style={{ height: 7 }}></Text>
 
-          <View style={styles.filterView}>
-            <View style={[styles.filterLabel]}>
-              <View style={{ flexDirection: "row" }}>
-                <MaterialCommunityIcons
-                  name="cup"
-                  size={18}
-                  style={{ paddingTop: 9, marginRight: 5 }}
-                  color="black"
-                />
-                <Text style={[styles.spaceLeft, styles.TextSmall]}>Drinks</Text>
-              </View>
-              <View
-                style={[
-                  styles.filterTagsContainer,
-                  { flexDirection: "column" },
-                ]}
-              >
+            <View style={styles.filterView}>
+              <View style={[styles.filterLabel, { marginBottom: 35 }]}>
+                <View style={{ flexDirection: "row" }}>
+                  <FontAwesome5
+                    name="hamburger"
+                    size={18}
+                    style={{ paddingTop: 9, marginRight: 5 }}
+                    color="black"
+                  />
+                  <Text style={[styles.spaceLeft, styles.TextSmall]}>Foods</Text>
+                </View>
                 <View
-                  style={{
-                    flexDirection: "row",
-                    flexWrap: "wrap",
-                    justifyContent: "space-evenly",
-                  }}
+                  style={[
+                    styles.filterTagsContainer,
+                    { flexDirection: "column" },
+                  ]}
                 >
-                  {/* Mapping of the tags from the JSON to the components*/}
-                  {filters.drinks.map((tag, index) => (
-                    <View key={index}>
-                      <FilterTag color="#FF6347" title={tag} />
-                    </View>
-                  ))}
+                  <View
+                    style={{
+                      flexDirection: "row",
+                      flexWrap: "wrap",
+                      justifyContent: "space-evenly",
+                    }}
+                  >
+                    {/* Mapping of the tags from the JSON to the components*/}
+                    {filters.foods.map((tag, index) => (
+                      <View key={index}>
+                        <FilterTag color="#C41ED4" title={tag} />
+                      </View>
+                    ))}
+                  </View>
                 </View>
               </View>
             </View>
-          </View>
-          <Text style={{ height: 7 }}></Text>
-
-          <View style={styles.filterView}>
-            <View style={[styles.filterLabel]}>
-              <View style={{ flexDirection: "row" }}>
-                <FontAwesome5
-                  name="hamburger"
-                  size={18}
-                  style={{ paddingTop: 9, marginRight: 5 }}
-                  color="black"
-                />
-                <Text style={[styles.spaceLeft, styles.TextSmall]}>Foods</Text>
-              </View>
-              <View
-                style={[
-                  styles.filterTagsContainer,
-                  { flexDirection: "column" },
-                ]}
-              >
-                <View
-                  style={{
-                    flexDirection: "row",
-                    flexWrap: "wrap",
-                    justifyContent: "space-evenly",
-                  }}
-                >
-                  {/* Mapping of the tags from the JSON to the components*/}
-                  {filters.foods.map((tag, index) => (
-                    <View key={index}>
-                      <FilterTag color="#FF6347" title={tag} />
-                    </View>
-                  ))}
-                </View>
-              </View>
-            </View>
-          </View>
-
+          </ScrollView>
           <Text style={{ height: 7 }}></Text>
           <Pressable
             style={[styles.applyButton]}
